Restructure project and remove necessity for macroexpansion hook.
The code is restructured into the code which actually implements the
API into the package CL-ENVIRONMENTS.CLTL2 and the package containing
the shadowed CL symbols moved into CL-ENVIRONMENTS-CL retaining
CL-ENVIRONMENTS as a nickname.

The shadowed macros directly enclose their arguments in the code
walking macro reducing the need for the macro expansion hook.